Fix Modal close
